# docs/swagger.yaml
basePath: /
definitions:
  domain.Category:
    properties:
      created_at:
        type: string
      display_order:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      parent_id:
        type: integer
      products:
        items:
          $ref: '#/definitions/domain.Product'
        type: array
      updated_at:
        type: string
    type: object
  domain.Product:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.CreateCategoryResquest:
    properties:
      display_order:
        type: integer
      image_url:
        type: string
      name:
        type: string
      parent_id:
        type: integer
    type: object
  dto.CreateProductRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
    type: object
  dto.ErrorResponse:
    properties:
      message:
        example: Please provide valid inputs
        type: string
      reason:
        example: detalhes do erro se houver
        type: string
    type: object
  dto.GetCategoriesSuccess:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.Category'
        type: array
      message:
        example: Categories
        type: string
    type: object
  dto.GetCategorySuccess:
    properties:
      data:
        $ref: '#/definitions/domain.Category'
      message:
        example: Categories
        type: string
    type: object
...
  dto.UpdateStockRequest:
    properties:
      stock:
        type: integer
    type: object
...
host: localhost:3333
info:
  contact:
    email: contato@otavioteixeira.com
    name: Otávio Augusto
    url: http://github.com/T4vexx
  description: API de e-commerce feita em Go com Fiber
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ecommerce Go API
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Retorna uma lista com todas as categorias disponíveis
      produces:
      - application/json
      responses:
        "200":
          description: Retorna todas as categorias
          schema:
            $ref: '#/definitions/dto.GetCategoriesSuccess'
        "404":
          description: Retorna a mensagem de erro
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtém todas as categorias
      tags:
      - Catalog
  /categories/{id}:
    get:
      description: Retorna os detalhes de uma categoria com base no ID informado
      parameters:
      - description: ID da Categoria
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Retorna a categoria por id caso o usuário tenha permissão
          schema:
            $ref: '#/definitions/dto.GetCategorySuccess'
        "404":
          description: Retorna a mensagem de erro
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtém categoria por ID
      tags:
      - Catalog
...
  /products:
    get:
      description: Retorna uma lista com todos os produtos disponíveis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Obtém todos os produtos
      tags:
      - Catalog
  /products/{id}:
    get:
      description: Retorna os detalhes de um produto com base no ID informado
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Obtém produto por ID
      tags:
      - Catalog
...
  /seller/categories:
    post:
      consumes:
      - application/json
      description: Cria uma categoria utilizando os dados fornecidos no corpo da requisição
      parameters:
      - description: Dados para criação da categoria
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryResquest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Cria uma nova categoria
      tags:
      - Catalog
  /seller/categories/{id}:
    delete:
      description: Remove uma categoria do sistema com base no ID informado
      parameters:
      - description: ID da Categoria
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Exclui uma categoria
      tags:
      - Catalog
    patch:
      consumes:
      - application/json
      description: Atualiza os dados de uma categoria existente com base no ID informado
      parameters:
      - description: ID da Categoria
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualização da categoria
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryResquest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Atualiza uma categoria
      tags:
      - Catalog
  /seller/products:
    get:
      description: Retorna a lista de produtos do vendedor autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Obtém os produtos do vendedor
      tags:
      - Catalog
    post:
      consumes:
      - application/json
      description: Cria um produto para o vendedor autenticado com os dados fornecidos
      parameters:
      - description: Dados para criação do produto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Cria um novo produto
      tags:
      - Catalog
  /seller/products/{id}:
    delete:
      description: Remove um produto do vendedor autenticado com base no ID informado
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Exclui um produto
      tags:
      - Catalog
    patch:
      consumes:
      - application/json
      description: Atualiza a quantidade em estoque de um produto do vendedor autenticado
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualização do estoque
        in: body
        name: stock
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Atualiza o estoque de um produto
      tags:
      - Catalog
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um produto do vendedor autenticado com base
        no ID informado
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados do produto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Edita um produto
      tags:
      - Catalog
...
swagger: "2.0"